requirements.txt
numpy
pandas
matplotlib
Pillow
typing
torch
